Move REDCap functions to utils
